FEAT: Add config editor
